LocalOCR_Application_SmolDocling/requirements.txt
streamlit
torch
transformers
pillow
docling